:truck: Moved existing CLI tests from test_cli into the tests/cli directory
