+++ tests/cli/test_train_cli.py
@@ -7,6 +7,8 @@ from typer.testing import CliRunner
 from icenet_mp.cli.main import app
 from icenet_mp.model_service import ModelService
 
+from .conftest import CustomCliRunner
+
 
 class FakeModelService:
     def __init__(self) -> None:
@@ -42,6 +44,20 @@ class TestTrainCLI:
         assert len(captured) == 1
         assert captured[0].model.name == "quick-test"
         assert service.calls == [(None, False)]
+
+    def test_help(self, runner: CustomCliRunner) -> None:
+        runner.check_output(
+            ["train", "--help"],
+            expected_patterns=[
+                r"Usage: imp train \[OPTIONS\] \[overrides\]...",
+                r"Train a model",
+                r"overrides\s+<str>\s+One or more space-separated Hydra config overrides",
+                r"--checkpoint-dir\s+<str>\s+Path to a directory of existing",
+                r"--config-name\s+<str>\s+Name of a file to load from the config",
+                r"--help\s+-h\s+Show this message and exit.",
+                r"--multistage\s+Train an EncodeProcessDecode model in",
+            ],
+        )
 
     def test_multistage_training_resolves_checkpoint_directory(
         self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
